remove nonNull because it looks have some bugs

+++ graphql/src/graphql/modelManager/ChatRoomMessage/mutation.ts
@@ -1,4 +1,4 @@
-import { extendType, nonNull, stringArg, intArg } from 'nexus'
+import { extendType, stringArg, intArg } from 'nexus'
 import { toPrismaId } from '@skeet-framework/utils'
 import { ChatRoomMessage } from 'nexus-prisma'
 import { CurrentUser } from '@/index'
@@ -9,23 +9,23 @@ export const ChatRoomMessageMutation = extendType({
     t.field('createChatRoomMessage', {
       type: ChatRoomMessage.$name,
       args: {
-        role: nonNull(stringArg()),
-        content: nonNull(stringArg()),
-        chatRoomId: nonNull(stringArg()),
+        role: stringArg(),
+        content: stringArg(),
+        chatRoomId: stringArg(),
       },
-      async resolve(_, args, ctx) {
-        const { role, content, chatRoomId } = args
-        const user: CurrentUser = ctx.user
-        console.log(user)
-        if (user.uid === '') throw new Error(`You are not logged in!`)
-
-        const data = {
-          role,
-          content,
-          userId: toPrismaId(user.id),
-          chatRoomId: toPrismaId(chatRoomId),
-        }
+      async resolve(_, { role, content, chatRoomId }, ctx) {
         try {
+          const user: CurrentUser = ctx.user
+          console.log(user)
+          if (user.uid === '') throw new Error(`You are not logged in!`)
+          if (!role || !content || !chatRoomId)
+            throw new Error(`not enough args`)
+          const data = {
+            role,
+            content,
+            userId: toPrismaId(user.id),
+            chatRoomId: toPrismaId(chatRoomId),
+          }
           return await ctx.prisma.chatRoomMessage.create({
             data,
           })
@@ -38,16 +38,17 @@ export const ChatRoomMessageMutation = extendType({
     t.field('updateChatRoomMessage', {
       type: ChatRoomMessage.$name,
       args: {
-        id: nonNull(stringArg()),
+        id: stringArg(),
         content: stringArg(),
         userId: intArg(),
         chatRoomId: intArg(),
       },
       async resolve(_, args, ctx) {
-        const id = toPrismaId(args.id)
-        let data = JSON.parse(JSON.stringify(args))
-        delete data.id
         try {
+          if (!args.id) throw new Error(`no id`)
+          const id = toPrismaId(args.id)
+          const data = JSON.parse(JSON.stringify(args))
+          delete data.id
           return await ctx.prisma.chatRoomMessage.update({
             where: {
               id,
@@ -63,10 +64,11 @@ export const ChatRoomMessageMutation = extendType({
     t.field('deleteChatRoomMessage', {
       type: ChatRoomMessage.$name,
       args: {
-        id: nonNull(stringArg()),
+        id: stringArg(),
       },
       async resolve(_, { id }, ctx) {
         try {
+          if (!id) throw new Error(`no id`)
           return await ctx.prisma.chatRoomMessage.delete({
             where: {
               id: toPrismaId(id),
